Move GltfMixin url/file handling to GltfCatalogitem
